+++ src/src_f/cmod_HelmholtzOp.f90
@@ -56,11 +56,11 @@ contains
         real(c_double), intent(inout):: lap (bL(1):(N(1)+bU(1)),bL(2):(N(2)+bU(2)),bL(3):(N(3)+bU(3)))
 
   
-        integer                ::  i, ii
-        integer                ::  j, jj
-        integer                ::  k, kk
+        integer(c_int)               ::  i, ii
+        integer(c_int)               ::  j, jj
+        integer(c_int)               ::  k, kk
   
-        real                   ::  dd1
+        real(c_double)                  ::  dd1
   
 
         !===========================================================================================================
